change to stratum protocol (difficulty)

+++ src/net/SubmitResult.cpp
@@ -29,7 +29,7 @@
 #include "net/SubmitResult.h"
 
 
-SubmitResult::SubmitResult(int64_t seq, uint32_t diff, uint64_t actualDiff) :
+SubmitResult::SubmitResult(int64_t seq, uint64_t diff, uint64_t actualDiff) :
     seq(seq),
     diff(diff),
     actualDiff(actualDiff),
